requirements.txt
# Django and Core Packages
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.6.0

# Database
psycopg2-binary==2.9.10

# Authentication
djangorestframework-simplejwt==5.4.0

# File Storage
django-storages==1.14.6
boto3==1.35.90
Pillow==11.1.0

# Admin Interface
django-admin-interface==0.30.1
django-colorfield==0.12.0

# Import/Export
django-import-export==4.3.3

# Filtering and Search
django-filter==24.3

# Background Tasks
celery==5.4.0
redis==5.2.1

# Configuration
python-decouple==3.8
dj-database-url==2.2.0

# Cache
django-redis==5.4.0

# Production Server
gunicorn==23.0.0
whitenoise==6.8.2

# Error Tracking (Optional)
sentry-sdk==2.20.0

# Development Tools
django-extensions==3.2.3

# Admin theme
django-unfold==0.63.0